Make rand a dev-dependency
