+++ Cargo.toml
@@ -17,11 +17,11 @@ default-features = false
 version = "2"
 default-features = false
 
-[dependencies.rand]
+[dev-dependencies.rand]
 version = "0.6"
 default-features = false
 
-[dependencies.rand_xorshift]
+[dev-dependencies.rand_xorshift]
 version = "0.1"
 default-features = false
